# Speed up initial refinement during mesh generation