Fixes #496.

Timing comparison (test code and results for `main` from #496):

Level | `main` | this branch
--: | --: | --:
9 | 46.2 s | 277 ms
10 | 799 s | 1.36 s